Add an app to preview components
This commit adds the ability to run an app similar to what you get if
you run `npm start` in the root directory of a full checkout of
govuk-frontend.

In fact, currently the easiest way to run this app is to run

    ./scripts/get-govuk-frontend.sh --extract-all
    ./scripts/run-app.sh

which leverages the gulp tasks of govuk-frontend to build needed assets.