Replace multiple string replacements with one generic one
We currently add a check if and when we encounter a `for attribute, value in` iteration.

This PR simplifies this a bit.

I've left the `(foo|bar)` structure to make it explicit what we've added. However, even
in those cases we have not explicitly checked ALL uses of that change.

This may be _good enough_ until we deprecate this repo.